AssertIsDefined util added. Auth middleware added. Specific user notes authentication added. User UI css typo fixed.

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticadUserId = req.session.userId;
+
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticadUserId);
+
+        const notes = await NoteModel.find({userId: authenticadUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -14,8 +19,12 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticadUserId = req.session.userId;
+        assertIsDefined(authenticadUserId);
     
     try {
+        assertIsDefined(authenticadUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -24,6 +33,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticadUserId)) {
+            throw createHttpError(401, "You cannot acces this note");
         }
 
         res.status(200).json(note);
@@ -40,13 +53,17 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticadUserId = req.session.userId;
+
 
     try {
+        assertIsDefined(authenticadUserId);
         if (!title) {
             throw createHttpError(400, "Note must have a title");
         }
 
        const newNote = await NoteModel.create({
+            userId: authenticadUserId,
             title: title,
             text: text,
        });
@@ -70,8 +87,12 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticadUserId = req.session.userId;
+
     
     try {
+        assertIsDefined(authenticadUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -84,6 +105,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticadUserId)) {
+            throw createHttpError(401, "You cannot acces this note");
         }
 
         note.title = newTitle;
@@ -99,8 +124,12 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async(req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticadUserId = req.session.userId;
+
 
     try {
+        assertIsDefined(authenticadUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -109,6 +138,10 @@ export const deleteNote: RequestHandler = async(req, res, next) => {
 
         if(!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticadUserId)) {
+            throw createHttpError(401, "You cannot acces this note");
         }
 
         await note.deleteOne();
